requirements.txt
pyqt5
pyserial
cryptography
pylint
pyinstaller
